requirements.txt
django==3.2
flake8==5.0.4
python-decouple==3.6
django-bootstrap-v5==1.0.11